+++ packages/expo-constants/CHANGELOG.md
@@ -11,6 +11,8 @@
 
 ### 🐛 Bug fixes
 
+- If `manifest` is undefined, try to grab it from `expo-updates` ([#10668](https://github.com/expo/expo/pull/10668) by [@esamelson](https://github.com/esamelson))
+
 ## 9.2.0 — 2020-08-18
 
 _This version does not introduce any user-facing changes._